Move compile-error-state orchestration from compile.py to server.py
The orchestration that wraps coqc-only `run_compile`/`run_compile_file`
with best-effort PET state capture was living in `compile.py`, which
forced that module to import from `interactive.py`.  This contradicted
its top-of-file invariant ("no pytanque dependency for core operation")
and meant a no-pet deployment paid a real `_run_with_pet` round-trip on
every compile failure before the error got swallowed.

Restore the invariant by moving the orchestration to `server.py` (where
the `@mcp.tool` wrappers live) and importing the pure helper
`_first_error_from_positions` from compile and `capture_position_state`
from interactive directly.  No public-API change: the failure dict shape
stays identical and the `run_compile_with_state` /
`run_compile_file_with_state` entry points keep their names (just on a
different module).

Tests that were monkeypatching `_compile._capture_compile_error_state`
now monkeypatch `_server._capture_compile_error_state`.  The full suite
still passes (732/732).
